requirements.txt
asgiref==3.4.1
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.7
django-utils-six==2.0
djangorestframework==3.12.4
pycparser==2.17
python-decouple==3.4
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
